continue organizing application
